添加由图片确定Token并求解的输入方式
1. 图片解析采用神经网络进行字符识别

+++ TokenExtract/picSlice.m
@@ -43,7 +43,7 @@ XSpan = arrayfun(@(x) round(indC(x)-intvC/2/2):round(indC(x)+intvC/2/2), 1:width
 for iter = 1:height * width
     ImgSet(:, :, iter) = 256 - imresize(rgb2gray(Img(YSpan{Y(iter)}, XSpan{X(iter)}, :)), [21 21]);
 end
-ImgMat = reshape(mean(ImgSet,[1 2]), [height width]) > 32;
+ImgMat = double(reshape(mean(ImgSet,[1 2]), [height width]) > 32);
 ImgSet(:,:,~ImgMat) = [];
